// apps/inatrace-fe/src/api/model/apiProcessingOrder.ts
export interface ApiProcessingOrder {
  /**
   * Entity id
   */
  id?: number;
  /**
   * Initiator user ID
   */
  initiatorUserId?: number;
  /**
   * Timestamp indicates when processing order have been created
   */
  creationTimestamp?: Date;
  processingAction?: ApiProcessingAction;
  /**
   * Processing date
   */
  processingDate?: string;
  /**
   * Input transactions
   */
  inputTransactions?: Array<ApiTransaction>;
  /**
   * Target stock orders
   */
  targetStockOrders?: Array<ApiStockOrder>;
}

/**
 * Namespace for property- and property-value-enumerations of ApiProcessingOrder.
 */
export namespace ApiProcessingOrder {
  /**
   * All properties of ApiProcessingOrder.
   */
  export enum Properties {
    /**
     * Entity id
     */
    id = 'id',
    /**
     * Initiator user ID
     */
    initiatorUserId = 'initiatorUserId',
    /**
     * Timestamp indicates when processing order have been created
     */
    creationTimestamp = 'creationTimestamp',
    processingAction = 'processingAction',
    /**
     * Processing date
     */
    processingDate = 'processingDate',
    /**
     * Input transactions
     */
    inputTransactions = 'inputTransactions',
    /**
     * Target stock orders
     */
    targetStockOrders = 'targetStockOrders',
  }

  export function formMetadata() {
    return {
      metadata: formMetadata,
      classname: 'ApiProcessingOrder',
      vars: [
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'id',
          classname: 'ApiProcessingOrder',
          dataType: 'number',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'initiatorUserId',
          classname: 'ApiProcessingOrder',
          dataType: 'number',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'creationTimestamp',
          classname: 'ApiProcessingOrder',
          dataType: 'Date',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          metadata: ApiProcessingAction.formMetadata,
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'processingAction',
          classname: 'ApiProcessingOrder',
          dataType: 'ApiProcessingAction',
          isPrimitiveType: false,
          isListContainer: false,
          complexType: 'ApiProcessingAction',
        },
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'processingDate',
          classname: 'ApiProcessingOrder',
          dataType: 'string',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          metadata: ApiTransaction.formMetadata,
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'inputTransactions',
          classname: 'ApiProcessingOrder',
          dataType: 'Array&lt;ApiTransaction&gt;',
          isPrimitiveType: false,
          isListContainer: true,
          complexType: 'ApiTransaction',
        },
        {
          metadata: ApiStockOrder.formMetadata,
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'targetStockOrders',
          classname: 'ApiProcessingOrder',
          dataType: 'Array&lt;ApiStockOrder&gt;',
          isPrimitiveType: false,
          isListContainer: true,
          complexType: 'ApiStockOrder',
        },
      ],
      validators: {
        id: [],
        initiatorUserId: [],
        creationTimestamp: [],
        processingAction: [],
        processingDate: [],
        inputTransactions: [],
        targetStockOrders: [],
      },
    };
  }

  // export const ApiProcessingOrderValidationScheme = {
  //     validators: [],
  //     fields: {
  //               id: {
  //                   validators: []
  //               },
  //               initiatorUserId: {
  //                   validators: []
  //               },
  //               creationTimestamp: {
  //                   validators: []
  //               },
  //               processingAction: {
  //                   validators: []
  //               },
  //               processingDate: {
  //                   validators: []
  //               },
  //               inputTransactions: {
  //                   validators: []
  //               },
  //               targetStockOrders: {
  //                   validators: []
  //               },
  //     }
  // } as SimpleValidationScheme<ApiProcessingOrder>;
}